* commit '0be43dfe3ad0ae07e895021bd24552e26875f6b7':
  Match Chart version
  Ensure .version advances in RB.
  Bump versions missed during ff-merge.
  Add comment about why a sleep is healthy during retry block.
  Rev chart version post merge.
  CASMCMS-6622: Re-initialize vault connection on 403 failure
  CASMCMS-6750: Fix unbound local variable during cleanup block

+++ src/cfsssh/vault.py
@@ -275,6 +275,7 @@ class VaultSshKey(object):
         """
         The public half of this key.
         """
+        private_key_path = None
         try:
             private_key_path = self.export_private_key()
             proc = subprocess.Popen(['ssh-keygen', '-y', '-f', private_key_path], stdout=subprocess.PIPE,
@@ -282,7 +283,7 @@ class VaultSshKey(object):
             out, _ = proc.communicate()
             return out.decode('utf-8')
         finally:
-            if os.path.exists(private_key_path):
+            if private_key_path and os.path.exists(private_key_path):
                 os.unlink(private_key_path)
 
     def export_public_key(self, destination=None):
